Refactor Linear client into package
